+++ example/lib/pages/connect.dart
@@ -45,7 +45,11 @@ class _ConnectPageState extends State<ConnectPage> {
     super.initState();
     _readPrefs();
     if (lkPlatformIs(PlatformType.android)) {
-      _checkPremissions();
+      _checkPermissions();
+    }
+
+    if (lkPlatformIsMobile()) {
+      LiveKitClient.initialize(bypassVoiceProcessing: true);
     }
   }
 
@@ -56,7 +60,7 @@ class _ConnectPageState extends State<ConnectPage> {
     super.dispose();
   }
 
-  Future<void> _checkPremissions() async {
+  Future<void> _checkPermissions() async {
     var status = await Permission.bluetooth.request();
     if (status.isPermanentlyDenied) {
       print('Bluetooth Permission disabled');